% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_job_status.R
\name{check_job_status}
\alias{check_job_status}
\title{Check the status of a job running on an HPC or locally}
\usage{
check_job_status(
  job_ids = NULL,
  user = NULL,
  scheduler_name = get_supported_schedulers(),
  standardize_output = TRUE,
  control = NULL
)
}
\arguments{
\item{job_ids}{A numeric or character vector of job IDs (Slurm/TORQUE) or
process IDS (local) to query.}

\item{user}{Optional character vector of username(s) used to filter processes
by owner.}

\item{scheduler_name}{Which scheduler to use for status query. Options are
'torque' and 'qsub' (aliases), 'slurm' and 'sbatch' (aliases), and 'local'
or 'sh' (aliases).}

\item{standardize_output}{Logical. If \code{TRUE}, will return standardized
values in the status column (e.g., "complete", "failed", etc.). Otherwise,
returns raw statuses depending on the scheduler. Default: \code{TRUE}.}

\item{control}{A list of additional, scheduler-specific arguments. See
Details.}
}
\value{
A data frame with the status information of the queried job(s).
Columns may vary depending on \code{columns} argument specification in
\code{control}.
}
\description{
\code{check_job_status} returns a data frame with requested job
information including status.
}
\details{
{
\code{check_job_status} is useful for checking the status of local jobs
launched via background processes or system calls, particularly in scripting 
or pipeline execution.

\code{check_job_status(scheduler = "slurm")} submits via \code{sacct}.
\code{check_job_status(scheduler = "torque")} submits via \code{qselect}.

\code{check_job_status(scheduler = "local")} is platform-dependent and
intended for UNIX-like systems, submiting via \code{ps}. If a PID does not
appear in the \code{ps} output, it is assumed to have completed and its
\code{STAT} value will be set to\code{"C"}. For running or sleeping
processes, \code{STAT} reflects the current process reported by the OS.

Columns of the output names are harmonized across systems (e.g., renaming
\code{S} to \code{STAT} and \code{COMMAND} to \code{COMM} if present). It
uses [data.table::fread()] to parse output.

The \code{control} argument can include additional, scheduler-specific
arguments as a named list:
\describe{
   \item{\code{columns}}{
     A character vector of scheduler-specific columns to include in
     the output. The default for Slurm is \code{c("jobid", "submit",
     "timelimit", "start", "end", "state")} and the default for local is
     \code{c("user", "pid", "state", "time", "etime", "\%cpu", "\%mem",
     "comm", "xstat")}
   }
 }
}
}
\examples{
\dontshow{if (interactive() && nzchar(Sys.which("sacct"))) withAutoprint(\{ # examplesIf}
check_job_status(scheduler = "slurm", job_ids = "16346781")
\dontshow{\}) # examplesIf}
}
