% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_methods.R
\docType{methods}
\name{clone}
\alias{clone}
\alias{clone.class_job}
\alias{clone.class_job_sequence}
\title{Clone a job or job sequence}
\usage{
clone(x, new_name = NULL)

## S7 method for class <hpcR::class_job>
clone(x, new_name = NULL)

## S7 method for class <hpcR::class_job_sequence>
clone(x, new_name = NULL)
}
\arguments{
\item{x}{An \link{hpcR} job or job sequence object.}

\item{new_name}{Optional. A new name for the cloned job or job sequence.}
}
\description{
Clone a job or job sequence
}
