% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_job.R
\docType{methods}
\name{compile_job}
\alias{compile_job}
\title{Prepare a job object for submission}
\usage{
compile_job(job, ...)
}
\arguments{
\item{job}{A job object to be compiled for submission}

\item{...}{Not currently used}
}
\description{
Prepare a job object for submission
}
