% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_methods.R
\docType{methods}
\name{format_print}
\alias{format_print}
\alias{format}
\alias{print}
\alias{format.class_job}
\alias{print.class_job}
\alias{format.class_job_summary}
\alias{print.class_job_summary}
\title{Format and print hpcR objects}
\arguments{
\item{x}{An hpcR object to be formatted or printed.}

\item{...}{Currently not used.}
}
\description{
Format and print hpcR objects
}
\keyword{internal}
