% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submit.R
\name{get_name_from_path}
\alias{get_name_from_path}
\title{Get the file name from a full path.}
\usage{
get_name_from_path(path)
}
\arguments{
\item{path}{character a file path which can be a full path or not.}
}
\value{
character the file name of the given full path
}
\description{
Such as the file name of folder/file.txt is file.txt.
}
