% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{helper_functions}
\alias{helper_functions}
\alias{format_wall_time}
\title{Convert time components to standardized wall time format for a scheduler}
\usage{
format_wall_time(days = 0, hours = 0, minutes = 0, seconds = 0)
}
\arguments{
\item{days}{Numeric. Number of days to convert to wall time format.}

\item{hours}{Numeric. Number of hours to convert to wall time format.}

\item{minutes}{Numeric. Number of minutes to convert to wall time format.}

\item{seconds}{Numeric. Number of seconds to convert to wall time format.}
}
\value{
Character string in format "MM[:SS]", "HH:MM:SS" or "D-HH[:MM][:SS]"
representing the total time specified by the input arguments.
}
\description{
This function takes numeric inputs for days, hours, minutes, and seconds, and
converts them to a standardized wall time format that can be used in HPC job
schedulers. The output format will be "MM[:SS]", "HH:MM:SS" or
"D-HH[:MM][:SS]", depending on the values of the input.
}
