% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_methods.R
\docType{methods}
\name{summary}
\alias{summary}
\alias{summary.class_job}
\title{Summarize a job object}
\usage{
## S7 method for class <hpcR::class_job>
summary(object, ...)
}
\arguments{
\item{object}{A job object to be summarized.}

\item{...}{Currently not used.}
}
\description{
Summarize a job object
}
\examples{
tmp_script <- tempfile(pattern = "my_script", fileext = ".R")
writeLines("print('ok')", tmp_script)

my_job <- rjob("my_job") +
  script(tmp_script) +
  scheduler("slurm") +
  resources(
    n_nodes = 2,
    n_cores = 4,
    wall_time = "01:00:00"
  )

summary(my_job)

}
