% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_job.R
\name{update_job}
\alias{update_job}
\alias{+.hpcR::class_job}
\alias{update.hpcR::class_job}
\title{Incrementally build a job using `+`}
\usage{
\method{+}{`hpcR::class_job`}(e1, e2)

\method{update}{`hpcR::class_job`}(e1, e2)
}
\description{
Incrementally build a job using `+`
}
\keyword{internal}
