% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\docType{methods}
\name{validate}
\alias{validate}
\title{Validate a job object}
\usage{
validate(x, stage = c("update", "submit"))
}
\arguments{
\item{x}{An \code{hpcR} object to be validated.}

\item{stage}{The stage of validation, either "update" (i.e., when the job is
being built) or "submit" (i.e., when the job is being submitted). Default:
"update".}
}
\description{
Validate a job object
}
