% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar.R
\name{build_job}
\alias{build_job}
\alias{rjob}
\alias{job_name}
\alias{script}
\alias{job_directory}
\alias{scheduler}
\alias{resources}
\alias{packages}
\alias{settings}
\title{Build a job to submit to an HPC}
\usage{
rjob(job_name = NULL)

job_name(job_name = NULL)

script(script_path = NULL)

job_directory(job_directory = NULL)

scheduler(scheduler_name = NULL)

resources(
  n_nodes = NULL,
  n_cores = NULL,
  wall_time = NULL,
  total_memory = NULL,
  memory_per_core = NULL
)

packages(package_names = NULL, installation_path = .libPaths())

settings(print_session_info = NULL, print_environment = NULL)
}
\arguments{
\item{job_name}{A character string. The name of the job used by the scheduler
and in the job output.}

\item{script_path}{A character string. The path to the script to be executed
by the HPC.}

\item{job_directory}{A character string. The path to the 'home' directory
for this job.}

\item{scheduler_name}{A character string. The scheduler to be used for this
job. Options are 'slurm' (or 'sbatch'), 'torque' (or 'qsub'), or 'local'
(or 'sh').}

\item{n_nodes}{A number or character string. The number of compute nodes
to be requested on the scheduler.}

\item{n_cores}{A number or character string. The number of CPUs to be
requested on the scheduler.}

\item{wall_time}{A number or character string. The compute time requested
on the cluster using format "MM[:SS]", "HH:MM:SS" or "D-HH[:MM][:SS]". The
\code{\link{format_wall_time}} function is helpful for converting time
components to the correct format.}

\item{total_memory}{A number or character string. The total amount of memory
to be requested by the job. This is mutually exclusive with
\code{memory_per_core}.}

\item{memory_per_core}{A number or character string. The amount of memory to
be requested per core. This is mutually exclusive with \code{total_memory}.}

\item{package_names}{A character vector. The names of packages to be loaded
for the job.}

\item{installation_path}{A character vector. The path(s) to the library
where packages are stored. Defaults to \code{.libPaths()}, which should
work most of the time.}

\item{print_session_info}{If \code{TRUE}, print the \code{sessionInfo()} and
\code{Sys.info()} in the output file when the job starts. Useful for
debugging problems with the compute environment or R installation. Default:
\code{FALSE}.}

\item{print_environment}{If \code{TRUE}, print the \code{Sys.getenv()} in the
output file when the job starts. This can produce a lot of output, but can be
useful if certain environment variables are not being found when your job
runs, leading it to fail. Default: \code{FALSE}.}
}
\value{
A job object with the properties specified by the user.
}
\description{
\pkg{hpcR} provides a dynamic, user-friendly way to build
jobs to submit to high-performance clusters (HPCs). Properties of the job
are added using the \code{+} operator (think \pkg{ggplot2})

Properties are represented as functions where the name of the function is the
property (e.g., \code{scheduler()}), or the group of properties (e.g.,
\code{resources(n_nodes = ...)}), and the arguments are the values. These
functions are light-weight and meant to be easy to remember. Additional
properties can always be added later.

Once a job is built, it can be submitted to the specified scheduler using the
\code{submit()} function. See \link{hpcR_methods} for more details on working
with job objects.
}
\details{
Building a job always starts by initializing the job object using a
language-specific function (e.g., \code{rjob()}), depending on the language
of your script/code. Properties are then successively added, and validated,
i.e., errors are thrown about formatting and other basic mistakes.

Property statements are either named by the property they represent (e.g.,
\code{script} takes a path to a script to be executed), or by the general
group of properties when a collection of properties are related (e.g.,
\code{resources()} takes arguments \code{n_nodes}, \code{n_cores}, etc.)
}
\examples{
\dontrun{
my_job <- rjob("my_job") +
   script("path/to/script.R") +
   job_directory("path/to/job_directory") +
   scheduler("slurm") +
   resources(n_nodes = 2, n_cores = 4,
             wall_time = "01:00:00",
             total_memory = "16G") +
   packages(package_names = c("dplyr", "ggplot2"))
}
}
