% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar.R
\name{build_job}
\alias{build_job}
\alias{rjob}
\alias{script}
\alias{job_directory}
\alias{job_name}
\alias{resources}
\alias{scheduler}
\title{Build a job to submit to an HPC}
\usage{
rjob(job_name = NULL)

script(script_path = NULL)

job_directory(job_directory = NULL)

job_name(job_name = NULL)

resources(
  n_nodes = NULL,
  n_cores = NULL,
  wall_time = NULL,
  total_memory = NULL,
  memory_per_core = NULL
)

scheduler(scheduler_name = NULL)
}
\arguments{
\item{job_name}{A character string. The name of the job used in dependency
specification and job scheduler naming.}

\item{job_directory}{A character string. The path to the 'home' directory
for this job. TODO: Say what will be stored here}

\item{n_nodes}{A number or character string. The number of compute nodes
to be requested on the scheduler.}

\item{n_cores}{A number or character string. The number of CPUs to be
requested on the scheduler.}

\item{wall_time}{A number or character string. The compute time requested
on the cluster using format "MM[:SS]", "HH:MM:SS" or "D-HH[:MM][:SS]"}

\item{total_memory}{A number or character string. The total amount of memory
to requested by the job. This is mutually exclusive with
\code{memory_per_core}.}

\item{memory_per_core}{A number or character string. The amount of memory to
be requested per core. This is mutually exclusive with \code{total_memory}.}

\item{script}{A character string. The path to the script to be executed by
the HPC.}

\item{scheduler}{A character string. The scheduler to be used for this job.
Options are slurm' (or 'sbatch'), 'torque' (or 'qsub'), or 'local' (or 'sh').}

\item{...}{Additional arguments to be passed to the property validator(s).
These are intended for internal or expert use, so can be ignored by most
users.}
}
\value{
A job object with the properties
}
\description{
\pkg{hpcR} provides a dynamic, user-friendly way to build
jobs to submit to high-performance clusters (HPCs). Properties of the job
are added using the \code{+} operator (think \pkg{ggplot2})

Properties are represented as functions where the name of the function is the
property (e.g., \code{scheduler()}), or the group of properties (e.g.,
\code{resources(n_nodes = ...)}), and the arguments are the values. These
functions are light-weight and meant to be easy to remember.
}
\details{
Building a job always starts by initializing the job object using a
language-specific function (e.g., \code{rjob()}), depending on the language
of your script. Properties are then successively added, and softly validated,
i.e., warnings are thrown about formatting and other basic mistakes.

Property statements are either named by the property they represent (e.g.,
\code{script} takes a path to a script to be executed), or by the general
group of properties when a collection of properties are related (e.g.,
\code{resources()} takes arguments \code{n_nodes}, \code{n_cores}, etc.)
}
\keyword{internal}
