% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_job_status.R
\name{check_job_status}
\alias{check_job_status}
\alias{check_job_status_sbatch}
\alias{check_job_status_slurm}
\alias{check_job_status_qsub}
\alias{check_job_status_torque}
\alias{check_job_status_sh}
\alias{check_job_status_local}
\title{Query job status by ID or user}
\usage{
check_job_status(
  job_ids = NULL,
  user = NULL,
  scheduler = "slurm",
  standardize = TRUE,
  control = list()
)

check_job_status_sbatch(
  job_ids = NULL,
  user = NULL,
  scheduler = "slurm",
  standardize = TRUE,
  control = list()
)

check_job_status_slurm(
  job_ids = NULL,
  user = NULL,
  scheduler = "slurm",
  standardize = TRUE,
  control = list()
)

check_job_status_qsub(
  job_ids = NULL,
  user = NULL,
  scheduler = "torque",
  standardize = TRUE,
  control = list()
)

check_job_status_torque(
  job_ids = NULL,
  user = NULL,
  scheduler = "torque",
  standardize = TRUE,
  control = list()
)

check_job_status_sh(
  job_ids = NULL,
  user = NULL,
  scheduler = "local",
  standardize = TRUE,
  control = list()
)

check_job_status_local(
  job_ids = NULL,
  user = NULL,
  scheduler = "local",
  standardize = TRUE,
  control = list()
)
}
\arguments{
\item{job_ids}{A numeric or character vector of job IDs (Slurm/TORQUE) or process
IDS (local) to query.}

\item{user}{Optional character vector of username(s) used to filter processes
by owner.}

\item{scheduler}{Which scheduler to use for status query. Options are 'qsub'
and 'torque' (aliases), sbatch' and 'slurm' (aliases), and 'sh' or 'local'
(aliases). Default: 'slurm'}

\item{standardize}{Logical. If \code{TRUE}, will return standardized values in
the status column (e.g., "complete", "failed", etc.). Otherwise, returns raw
statuses depending on the scheduler. Default: TRUE}

\item{control}{A list of additional, scheduler-specific arguments. See Details.}
}
\value{
A data frame with the status information of the queried job(s). Columns
may vary depending on \code{columns} argument specification in \code{control}.
}
\description{
`check_job_status` returns a data frame with requested job information including status.
 Also available are scheduler-specific submit functions of the form `check_job_status_<scheduler>`
 where `<scheduler>` can be `slurm` or `sbatch` (aliases), `torque` or `qsub`
 (aliases), or `local` or `sh` (aliases).
}
\details{
{
`check_job_status` is useful for checking the status of local jobs launched via
background processes or system calls, particularly in scripting or pipeline
execution.

`check_job_status(scheduler = "slurm")` submits via `sacct`.
`check_job_status(scheduler = "torque")` submits via `qselect`.

\code{check_job_status(scheduler = "local")} is platform-dependent and intended
for UNIX-like systems, submiting via `ps`. If a PID does not appear in the `ps`
output, it is assumed to have completed and its `STAT` value will be set to
`"C"`. For running or sleeping processes, `STAT` reflects the current process
state reported by the OS.

Columns of the output names are harmonized across systems (e.g., renaming `S`
to `STAT` and `COMMAND` to `COMM` if present). It uses `fread()` to parse output.

The \code{control} argument can include additional, scheduler-specific arguments
as a named list:
\describe{
   \item{\code{columns}}{
     A character vector of scheduler-specific columns to include in
     the output. The default for Slurm is \code{c("jobid", "submit", "timelimit",
     "start", "end", "state")} and the default for local is \code{c("user",
     "pid", "state", "time", "etime", "\%cpu", "\%mem", "comm", "xstat")}
   }
}
}
}
\examples{
\dontrun{
  check_job_status(scheduler = "slurm", job_ids = "16346781")}
}
\author{
Michael Hallquist, Zach Vig
}
