% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{class_job}
\alias{class_job}
\alias{is_job}
\title{Top-level job object class}
\usage{
class_job(
  job_name = character(0),
  scheduler = class_pb_scheduler(),
  script = class_pb_script(),
  job_directory = character(0),
  resources = class_pb_resources(),
  .locked = FALSE
)

is_job(x, language = NULL)
}
\description{
The basis of job objects in \pkg{hpcR} is the \code{class_job}
class. This class stores all information needed to submit a job
to a high-performance cluster (HPC). Specific job types, such as
R jobs, are implemented as sub-classes of \code{class_job} and help with
specific property storage, job validation, and method dispatch.
}
\details{
Job objects are meant to be mostly for internal use, thus they
are intentionally difficult to edit using the usual S7 approach
(\code{S7_class@property <- value}). Instead, job objects have a
\code{.locked} property which, when \code{TRUE}, prevents editing. This
property is handled by the \code{+} operator approach to job building as
well as the \link{edit} approach to job editing. This ensures job objects
can be strictly validating and protects them against the vast majority
of possible (user) errors. See \link{build_job} for details about the
properties of a job object.
}
\keyword{internal}
