% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{class_job_update}
\alias{class_job_update}
\alias{is_job_update}
\title{Class for storing job updates}
\usage{
class_job_update(updates = list())

is_job_update(x)
}
\description{
The \code{class_job_update} class is the basis of all update
objects in \pkg{hpcR}. These objects are used to update job
objects when using the \code{+} operator. Job updates are lightweight
and usually store a few properties, or a single \link{PropertyBlock}.
}
\keyword{internal}
