% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{class_property_block}
\alias{class_property_block}
\alias{class_pb_resources}
\alias{class_pb_script}
\alias{class_pb_scheduler}
\alias{is_property_block}
\title{Class for storing related groups of job properties}
\usage{
class_property_block()

class_pb_resources(
  n_nodes = character(0),
  n_cores = character(0),
  wall_time = character(0),
  total_memory = character(0),
  memory_per_core = character(0)
)

class_pb_script(
  script_path = character(0),
  extension = character(0),
  language = character(0)
)

class_pb_scheduler(scheduler_name = character(0))

is_property_block(x, type = NULL)
}
\description{
The \code{PropertyBlock} class is an abstract
class for all property blocks in \pkg{hpcR}. Property
blocks are collections of related job properties, such as
resource specifications. These classes are intended to help
with quicker/easier validation.
}
\keyword{internal}
