% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_job.R
\docType{methods}
\name{compile_job}
\alias{compile_job}
\alias{compile}
\title{Prepare a job object for submission}
\usage{
compile(x)
}
\arguments{
\item{x}{The object to be compiled.}
}
\description{
Prepare a job object for submission
}
