% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpcR-package.R
\docType{package}
\name{hpcR-package}
\alias{hpcR}
\alias{hpcR-package}
\title{hpcR: Submit Jobs to High Performance Computing Clusters}
\description{
More about what it does (maybe more than one line). Continuation lines should be indented.
}
\author{
\strong{Maintainer}: Zach Vig \email{zachvig@unc.edu}

Authors:
\itemize{
  \item Michael Hallquist
}

}
\keyword{internal}
