% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpcR-package.R
\docType{package}
\name{hpcR-package}
\alias{hpcR}
\alias{hpcR-package}
\title{hpcR: Submit Jobs to High Performance Computing Clusters}
\description{
hpcR provides a simple interface for building jobs to submit to a high-performance cluster (HPC) and tracking them post-submission. For the more advanced user, the package also provides a straightforward but robust system for creating job dependencies/sequences for larger-scale projects.
}
\author{
\strong{Maintainer}: Zach Vig \email{zachvig@unc.edu}

Authors:
\itemize{
  \item Zach Vig \email{zachvig@unc.edu}
  \item Michael Hallquist
}

}
\keyword{internal}
