% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{hpcR_classes}
\alias{hpcR_classes}
\title{Classes for constructing jobs}
\description{
\pkg{hpcR} internally uses a suite of S7 classes
to construct and manage job objects. These allow for strict
validation and method dispatch. Technically a user could interact
directly with these classes, but they're optimized for the user-friendly
\code{+} operator approach to building jobs.
}
\section{Top-Level S7 Classes}{


Job objects store information about jobs intended to be submitted
to high-performance clusters (HPCs). The following are top-level
\link[S7]{S7_class} objects used in constructing, updating, and storing
properties of job objects.
\itemize{
   \item{\code{\link{class_job}}:}{
     Top-level class for all jobs. Basis of the "job object" that
     stores all information needed to submit a job.
   }
   \item{\code{\link{class_job_update}}:}{
     Top-level class for all updates. These are intermediate
     objects used to update job objects when using the \code{+} operator.
   }
   \item{\code{\link{class_property_block}}:}{
     Top-level (abstract) class for all property blocks, which are
     collections of related job properties. These are useful for
     validation and method dispatch.
   }
}
}

\section{S7 Subclasses}{


For specific job types and property blocks, \pkg{hpcR} uses
a collection of S7 subclasses. These include:

\itemize{
   \item{\code{\link{class_pb_resources}}}{
     Subclass of \code{class_property_block} for storing
     resource-related job properties, such as number of nodes,
     number of cores, wall time, and memory.
   }
   \item{\code{\link{class_pb_script}}}{
     Subclass of \code{class_property_block} for storing
     script-related job properties, such as script path,
     extension, and language.
   }
}
}

\section{Rationale}{

The central idea of job construction is that job objects
are incrementally updated such that \code{class_job <- class_job +
class_job_update}. This keeps the internals lightweight (e.g., as opposed to
creating two job objects and combining them). Properties can be stored
directly in the job object, e.g., \code{class_job@property}, or in the more
formal \code{class_property_block} class. Property blocks are themselves
collections of properties, e.g., \code{class_property_block@property}, but
they make it easier to do validation on properties that interact (e.g.,
mutually exclusive properties). \code{PropertyBlock} is an abstract class,
thus it's purpose is for specific sub-classes to inherit it.
}

\keyword{internal}
