% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_methods.R
\docType{methods}
\name{hpcR_methods}
\alias{hpcR_methods}
\alias{submit}
\title{Working with hpcR job objects}
\usage{
submit(job, ...)
}
\arguments{
\item{job}{A job object}

\item{...}{Not currently used}
}
\description{
Working with hpcR job objects
}
