% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_methods.R
\name{hpcR_methods}
\alias{hpcR_methods}
\title{Working with hpcR objects}
\description{
\pkg{hpcR} provides a set of methods for working with job and
job sequence objects. Methods include \code{submit()} for submitting jobs,
\code{clone()} for cloning, and \code{summary()} for summarizing job objects.
}
\section{Methods}{

\describe{
  \item{\link{submit}}{Submit a job or job sequence.}
  \item{\link{clone}}{Clone a job or job sequence.}
  \item{\link{summary}}{Create a summary of a job.}
  \item{\link{print}}{Print an hpcR object.}
}
}

\seealso{
\code{\link{submit}},
  \code{\link{clone}},
  \code{\link{summary}},
  \code{\link{print}}
}
