% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_job_status.R, R/utils.R
\name{get_default_status_columns}
\alias{get_default_status_columns}
\alias{hpcR_utils}
\alias{standardize_scheduler_name}
\alias{get_supported_schedulers}
\alias{get_supported_languages}
\title{Utility functions for hpcR}
\usage{
get_default_status_columns(scheduler_name = c("slurm", "torque", "local"))

standardize_scheduler_name(scheduler_name, strict = FALSE)

get_supported_schedulers(hpc_only = FALSE, include_alias = TRUE)

get_supported_languages()
}
\arguments{
\item{scheduler_name}{Scheduler name to standardize}

\item{strict}{Logical. If \code{TRUE}, throws an error if the scheduler name
is not recognized. Otherwise, returns the input scheduler name.}

\item{hpc_only}{Logical. If \code{TRUE}, only returns HPC schedulers,
excluding local scheduler and aliases. Default: \code{FALSE}.}

\item{include_alias}{Logical. If \code{TRUE}, includes aliases for HPC and 
local schedulers. Default: \code{FALSE}.}
}
\description{
Utility functions for hpcR
}
