% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invoke_system.R
\name{invoke_system}
\alias{invoke_system}
\title{Invoke R's `system` for local commands}
\usage{
invoke_system(
  input,
  is_script = TRUE,
  push_cmd = "sh",
  sched_args = NULL,
  echo = TRUE,
  fail_on_error = FALSE
)
}
\arguments{
\item{input}{A character string. Either the path to a script to execute or
a single line command.}

\item{push_cmd}{A character string. The command to call to push the script/command}

\item{sched_args}{A character vector of scheduler arguments.}

\item{echo}{Logical. If \code{TRUE}, prints final command to console.}

\item{fail_on_error}{If \code{TRUE}, fails when an error occurs. Otherwise,
just a warning is thrown.}

\item{is_script.}{Logical. Set to \code{TRUE} if the input is a script.}
}
\description{
Invoke R's `system` for local commands
}
\keyword{internal}
