% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequencing.R
\docType{methods}
\name{job_sequences}
\alias{job_sequences}
\alias{\%->\%}
\alias{job_sequence}
\alias{branch}
\title{Create job sequences using the \code{\%->\%} operator}
\usage{
lhs \%->\% rhs

job_sequence(sequence_name = NULL, ...)

branch(...)
}
\arguments{
\item{lhs}{A job sequence or a job object. This specifies which job or
jobs are "upstream" from those in \code{rhs}.}

\item{rhs}{A job sequence or job object. This specifies which job or jobs are
"downstream" from those in \code{lhs}. Note that \code{branch()} can be used
to specify multiple downstream jobs and returns a job sequence.}

\item{sequence_name}{A character string. The name of the job sequence to
create.}

\item{...}{Branching jobs or sequences to be added to a job sequence.}
}
\value{
A job sequence object
}
\description{
\pkg{hpcR} provides intuitive tools for building both simple and complex
sequences of jobs. Sequences are useful when one or more 'downstream' jobs
depend on one or more 'upstream' jobs.

Job schedulers have built-in functionality for managing job sequencing
(dependencies), but setting dependencies requires knowing job ID numbers
which are only available after submission. \pkg{hpcR} only requires the user
to build a sequence object and handles everything else internally
post-submission.

\pkg{hpcR} introduces the "arrow operator" (\code{\%->\%}) used to specify
the order of the sequence, e.g., \code{job_A \%->\% job_B}. Branching and
parallel downstream jobs are supported via the \code{branch()} function.
Alternative, specifying dependencies via the \code{sequencing()}  sugar
function is supported. See details below.
}
\details{
...
}
\examples{
job_A <- rjob("job_A")
job_B <- rjob("job_B")
job_C <- rjob("job_C")
job_D <- rjob("job_D")

# simple linear sequencing
sequence_1 <- job_A \%->\% job_B \%->\% job_C \%->\% job_D
# or equivalently
sequence_2 <- job_sequence("my_sequence") \%->\%
 job_A \%->\%
 job_B \%->\%
 job_C \%->\%
 job_D

# creating branches using `branch()`
sequence_3 <- job_sequence("my_branching_sequence") \%->\%
 job_A \%->\%
 branch(
   job_B,
   job_C \%->\% job_D
 )
# this sequence starts with A, then branches into parallel sequences,
# B and C -> D

}
