% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_job.R
\name{R_job}
\alias{R_job}
\title{R6 Class for R Jobs}
\description{
`R_job` is a user-friendly R6 class for creating jobs to submit to a high
performance computing (HPC) cluster using a scheduler.
}
\details{
...
}
\examples{
\dontrun{}

}
\author{
Michael Hallquist, Zach Vig
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{job_name}}{A user-defined name for the job used for specifying job
dependencies and informative job status queries on a job scheduler}

\item{\code{job_dir}}{Location of "home" directory for this job. Needs to be somewhere
that both compute nodes and login nodes can access, so /tmp is not suggested.}

\item{\code{scheduler}}{The job scheduler to be used for this batch. Options are:
"slurm"/"sbatch", "torque"/"qsub", or "local"/"sh".}

\item{\code{n_nodes}}{The number of nodes to be requested on the job scheduler}

\item{\code{n_cores}}{The number of cores (aka 'cpus', ignoring hyperthreading)
to be requested on the job scheduler}

\item{\code{wall_time}}{The amount of time requested on the job scheduler,
following d-hh:mm:ss format. Defaults to
   "4:00:00", which is 4 hours.}

\item{\code{mem_total}}{The total amount of memory (RAM) requested by the job}

\item{\code{mem_per_cpu}}{The amount of memory (RAM) requested per cpu
(total = mem_per_cpu * n_cores)}

\item{\code{array}}{A set of indexes specifying how many iterations of the job to
run and how to identify them, e.g., "1-10".}

\item{\code{parent_jobs}}{A vector of job names that are upstream of this job
and influence its execution}

\item{\code{wait_for_subs}}{If TRUE, code will be inserted to wait for all jobs
in a vector called \code{SUB_JOB_IDS} to finish before the batch exits.
It's up to your code to use this variable name}

\item{\code{all_subs_success}}{If TRUE, all jobs in vector \code{SUB_JOB_IDS}
must be successful for this job to finish (see \code{wait_for_subs} field)}

\item{\code{sqlite_db}}{File path to job tracking SQLite database}

\item{\code{scheduler_options}}{An optional character vector of scheduler arguments
to be included in the batch script header that control additional features
such as job emails or group permissions. These directives are added with #SBATCH
or #PBS headings, depending on the scheduler, and are ignored if the scheduler
is "local".}

\item{\code{r_packages}}{The R packages to be loaded into the environment before
job execution. These are loaded by pacman::p_load, which will install any
missing packages before attempting to load}

\item{\code{input_objects}}{An environment containing all objects to be written
to an RData object and passed to the batch job at execution}

\item{\code{input_rdata_file}}{The name of the environment to be loaded at the
beginning of the R batch prior to executing other code. Used to setup any
local objects needed to begin computation.}

\item{\code{shell_code}}{Shell code to be included in the batch script prior to
the R code to be run. This can include module load statements, environment
variable exports, etc.}

\item{\code{output_rdata_file}}{The name of the environment to be saved at the
end of the R batch execution, which can then be loaded by subsequent jobs.}

\item{\code{control.submit_job}}{A named list with control variables used by
\code{wait_for_job} when waiting for parent jobs to finish. Arguments include
\code{repolling_interval}, which is the number of seconds to wait between
successive checks on whether parent jobs have completed, and \code{max_wait}
which is the maximum number of seconds to wait for parent jobs to finish
before ending execution.}

\item{\code{control.wait_for_subs}}{A named list with control variables used by
\code{wait_for_job} when waiting for sub jobs to finish before this job
continues. See \code{control.submit_job} for details.}

\item{\code{print_session_info}}{If TRUE, print the `sessionInfo()` and `Sys.info()`
when the job starts. Useful for debugging problems with the compute
environment or R installation.}

\item{\code{print_environment}}{If `TRUE`, print the `Sys.getenv()` when the job
starts. This can produce a lot of output, but can be useful if certain
environment variables are not being found when your job runs, leading it to fail.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{r_script}}{The R code to be executed by this job. This can be a character
vector that includes multiple R statements or an expression object containing
the R code to be evaluated}

\item{\code{r_code}}{The R code to be executed by this job. This can be a character
vector that includes multiple R statements or an expression object containing
the R code to be evaluated. (Mutually exclusive with \code{r_script})}

\item{\code{post_subs_r_script}}{The path to an R script to be executed by the batch after
all sub jobs complete Only relevant if \code{wait_for_subs = TRUE}. (Mutually
exclusive with \code{post_subs_r_code}).}

\item{\code{post_subs_r_code}}{The R code to be executed by the batch after all sub jobs
complete. This can be a character vector that includes multiple R statements
or an expression object containing the R code to be evaluated. Only relevant
if \code{wait_for_subs = TRUE}. (Mutually exclusive with \code{post_subs_r_script})}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-R_job-new}{\code{R_job$new()}}
\item \href{#method-R_job-submit}{\code{R_job$submit()}}
\item \href{#method-R_job-copy}{\code{R_job$copy()}}
\item \href{#method-R_job-get_job_id}{\code{R_job$get_job_id()}}
\item \href{#method-R_job-clone}{\code{R_job$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-R_job-new"></a>}}
\if{latex}{\out{\hypertarget{method-R_job-new}{}}}
\subsection{Method \code{new()}}{
Create a new R_job object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R_job$new(
  r_script = NULL,
  r_code = NULL,
  job_name = NULL,
  job_dir = NULL,
  scheduler = NULL,
  n_nodes = NULL,
  n_cores = NULL,
  wall_time = NULL,
  mem_total = NULL,
  mem_per_cpu = NULL,
  array = NULL,
  parent_jobs = NULL,
  wait_for_subs = FALSE,
  all_subs_success = FALSE,
  post_subs_r_script = NULL,
  post_subs_r_code = NULL,
  sqlite_db = NULL,
  scheduler_options = NULL,
  r_packages = NULL,
  input_rdata_file = NULL,
  input_objects = NULL,
  shell_code = NULL,
  output_rdata_file = NULL,
  control.submit_job = NULL,
  control.wait_for_subs = NULL,
  print_session_info = TRUE,
  print_environment = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{r_script}}{The path to an R script to be executed by the batch.
(Mutually exclusive with \code{r_code})}

\item{\code{r_code}}{A character vector or expression containing R code to be executed.
(Mutually exclusive with \code{r_script})}

\item{\code{job_name}}{A character string. The name of the job used in dependency
specification and job scheduler naming.}

\item{\code{job_dir}}{A character string. The path to the 'home' directory for this job.
TODO: Say what will be stored here}

\item{\code{scheduler}}{A character string. The scheduler to be used for this compute.
Options are slurm' (or 'sbatch'), 'torque' (or 'qsub'), or 'local' (or 'sh').}

\item{\code{n_nodes}}{A number or character string. The number of compute nodes
to be requested on the scheduler.}

\item{\code{n_cores}}{A number or character string. The number of CPUs to be
requested on the scheduler.}

\item{\code{wall_time}}{A number or character string. The compute time requested
on the cluster using format "dd-HH:MM:SS".}

\item{\code{mem_total}}{A number or character string. The total amount of memory
to requested by the job.}

\item{\code{mem_per_cpu}}{A number or character string. The amount of memory to
be requested per CPU}

\item{\code{array}}{An integer vector or character string. If specified, will run
as many iterations as is the length of the vector, with each iteration being
identified by that number, e.g., \code{c(1, 2, 3)} or "1-3" will iterate
over the job three times. Schedulers also accept sequence notation, e.g.,
"1-13:3" iterates from 1 to 13 by intervals of 3 (i.e., 1, 4, 7, 10, 13).
The job array number will be made available to your R script as \code{ARRAY_ID}}

\item{\code{parent_jobs}}{Numberical or character vector of one or more job ids
that are parents of this job.}

\item{\code{wait_for_subs}}{Logical. If \code{TRUE}, do not end this job until all
sub-jobs have completed. Assign sub-job ids to \code{SUB_JOB_IDS} in your
code. Default: \code{FALSE}}

\item{\code{all_subs_success}}{If \code{TRUE}, don't count this job as successful
unless all sub-jobs are successful. Default: \code{FALSE}}

\item{\code{post_subs_r_script}}{The path to an R script to be executed by the batch
after sub jobs have completed. Only relevant if \code{wait_for_subs = TRUE}
(Mutually exclusive with \code{post_subs_r_code})}

\item{\code{post_subs_r_code}}{A character vector or expression containing R code to be
executed after sub jobs have completed. Only relevant if \code{wait_for_subs = TRUE}
(Mutually exclusive with \code{post_subs_r_script})}

\item{\code{sqlite_db}}{A character string. The location of the SQLite database
to be used for job tracking. If `NULL`, job tracking will be disabled.
TODO: If sqlite db provided doesn't exist, write to job_dir}

\item{\code{scheduler_options}}{A character vector of scheduler options to be
added to the header of the batch script}

\item{\code{r_packages}}{A character vector of R packages to be loaded when
compute script runs}

\item{\code{input_rdata_file}}{A character string specifying the path to the
environment to be loaded at the beginning of the R batch prior to executing code}

\item{\code{input_objects}}{A list or environment in the current execution environment
to be cached and used as input to the R batch. This is mutually exclusive
with input_rdata_file at present.}

\item{\code{shell_code}}{A character vector of shell code to be included in the batch
script that submits the job}

\item{\code{output_rdata_file}}{The name of the environment to be saved at the
end of the R batch execution}

\item{\code{control.submit_job}}{A named list with control variables used by
\code{submit_job} when submitting this job, e.g., \code{repolling_interval},
\code{max_wait}, and \code{wait_signal}. This argument defaults when
\code{scheduler = "local"}. For expert use only.}

\item{\code{control.wait_for_subs}}{A named list with control variables used by
\code{wait_for_job} when waiting for sub jobs to finish. Arguments include
\code{repolling_interval} and \code{max_wait}. This argument defaults when
\code{wait_for_subs = TRUE}. For expert use only.}

\item{\code{print_session_info}}{Logical. If \code{TRUE}, print information about
the R environment `sessionInfo()` and compute environment `Sys.info()` when
the job starts. Default: \code{TRUE}}

\item{\code{print_environment}}{Logical. If \code{TRUE}, print the session environment
via `Sys.getenv()` when the job starts. Default: \code{FALSE}.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
The scheduler-generated job ID is made available to your R
script as \code{JOB_ID}.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-R_job-submit"></a>}}
\if{latex}{\out{\hypertarget{method-R_job-submit}{}}}
\subsection{Method \code{submit()}}{
Submit job to scheduler or local compute
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R_job$submit()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-R_job-copy"></a>}}
\if{latex}{\out{\hypertarget{method-R_job-copy}{}}}
\subsection{Method \code{copy()}}{
Create a deep copy of a batch job with minor changes
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R_job$copy(
  r_script = NULL,
  r_code = NULL,
  job_name = NULL,
  n_nodes = NULL,
  n_cores = NULL,
  wall_time = NULL,
  mem_total = NULL,
  mem_per_cpu = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{r_script}}{The path to an R script to be executed by the batch.
(Mutually exclusive with \code{r_code})}

\item{\code{r_code}}{A character vector or expression containing R code to be executed.
(Mutually exclusive with \code{r_script})}

\item{\code{job_name}}{The name of the job used in dependency specification and job
scheduler naming}

\item{\code{n_nodes}}{A number or character string. The number of compute nodes
to be requested on the scheduler.}

\item{\code{n_cores}}{A number or character string. The number of CPUs to be
requested on the scheduler.}

\item{\code{wall_time}}{A number or character string. The compute time requested
on the cluster using format "dd-HH:MM:SS".}

\item{\code{mem_total}}{A number or character string. The total amount of memory
to requested by the job.}

\item{\code{mem_per_cpu}}{A number or character string. The amount of memory to
be requested per CPU}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
This method exposes a few named parameters that can be used to override
the copied fields with new values to avoid needing to change these one-by-one
using obj$<field> <- x syntax
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-R_job-get_job_id"></a>}}
\if{latex}{\out{\hypertarget{method-R_job-get_job_id}{}}}
\subsection{Method \code{get_job_id()}}{
Return the job id of this job (populated by job submission)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R_job$get_job_id()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-R_job-clone"></a>}}
\if{latex}{\out{\hypertarget{method-R_job-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R_job$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
