% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_methods.R
\docType{methods}
\name{submit}
\alias{submit}
\alias{submit.class_job}
\alias{submit.class_job_sequence}
\title{Submit a job or job sequence to an HPC scheduler}
\usage{
submit(x)

## S7 method for class <hpcR::class_job>
submit(x)

## S7 method for class <hpcR::class_job_sequence>
submit(x)
}
\arguments{
\item{x}{An \link{hpcR} job or job sequence object.}
}
\description{
Submit a job or job sequence to an HPC scheduler
}
\examples{
tmp_script <- tempfile(pattern = "my_script", fileext = ".R")
writeLines("print('ok')", tmp_script)

my_job <- rjob("my_job") +
  script(tmp_script) +
  scheduler("slurm") +
  resources(
    n_nodes = 2,
    n_cores = 4,
    wall_time = "01:00:00"
  )

\dontrun{
submit(my_job)
}
}
