% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submit_job.R
\name{submit_job}
\alias{submit_job}
\title{Submit a script or command to a high-performance cluster using a scheduler}
\usage{
submit_job(
  input,
  input_type = c("script", "oneliner"),
  scheduler_name = get_supported_schedulers(),
  fail_on_error = FALSE,
  depends_on = NULL,
  env_variables = NULL,
  echo = FALSE,
  control = list(),
  .call = rlang::caller_env()
)
}
\arguments{
\item{input}{A path to a script that should be executed by the scheduler or a
single command to be run. In the case of conflicts, the directives passed
with \code{scheduler_arguments} will take precedence. Required.}

\item{input_type}{Character. Either 'script' or 'oneliner' to indicate
whether the input is a script or a single line command. Default: 'script'}

\item{scheduler_name}{Which scheduler to use for job submission. Options are
'torque', 'slurm', 'local', 'sbatch', 'qsub', and 'sh'. The terms 'qsub' and
'torque' are aliases (where 'torque' submits via the qsub command). Likewise
for 'sbatch' and 'slurm'. The scheduler 'sh' (alias, 'local') does not submit
to any scheduler at all, but instead executes the command immediately via sh.
Default: 'slurm'}

\item{fail_on_error}{Whether to stop execution of the script (\code{TRUE}),
or issue a warning (\code{FALSE}) if the job submission fails. Default:
FALSE (i.e., issue a warning).}

\item{depends_on}{An optional character string of jobs or process ids that
should complete before this job is executed.}

\item{env_variables}{An optional named character vector containing
environment variables and their values to be passed to the \code{script} at
execution time. This is handled by the '-v' directive on TORQUE clusters and
by '--export' on Slurm clusters. The names of this vector are the environment
variable names and the values of the vector are the environment variable
values to be passed in. If you want to propagate the current value of an
environment variable to the compute node at runtime, use NA as the value of
the element in \code{env_variables}. See examples.}

\item{echo}{Whether to echo the job submission command to the terminal at the
time it is scheduled. Default: \code{TRUE}}

\item{control}{A list of additional, scheduler-specific arguments. See
details.}

\item{.call}{The calling environment, used for error messaging.}
}
\value{
A character string containing the job ID of the scheduled job.
}
\description{
\code{submit_job} submits a single script or a single command
to a high-performance cluster (HPC) using a specified scheduler (TORQUE,
Slurm) or locally.
}
\details{
The \code{control} argument can include additional, scheduler-
specific arguments as a named list:
\describe{
   \item{\code{export_all}}{
   (For TORQUE or Slurm schedulers) Whether to export all environment
   variables to the compute node at runtime. Default: FALSE
   }
   \item{\code{scheduler_arguments}}{
   An optional character vector of arguments to be included in
   the scheduling command. On TORQUE, these will typically begin with '-l'
   such as '-l wall_time=10:00:00'.
   }
   \item{\code{wait_signal}}{
   (For TORQUE or Slurm schedulers) The signal that should
   indicate that parent jobs have finished. Default: 'afterok'
   }
   \item{\code{repolling_interval}}{
   (For local jobs) The number of seconds to wait
   before rechecking job status. Default: 60
   }
   \item{\code{max_wait}}{
   How long to wait on the job before giving up, in seconds.
   Default: 24 hours (86,400 seconds)}
}
}
\examples{
\dontshow{if (interactive() && nzchar(Sys.which("sbatch"))) withAutoprint(\{ # examplesIf}
job_id <- submit_job(input = "my_script.R", scheduler_name = "slurm")
\dontshow{\}) # examplesIf}
}
