% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submit_job.R
\name{submit_job}
\alias{submit_job}
\alias{submit_cmd}
\alias{submit_job_sbatch}
\alias{submit_job_slurm}
\alias{submit_cmd_sbatch}
\alias{submit_cmd_slurm}
\alias{submit_job_qsub}
\alias{submit_job_torque}
\alias{submit_cmd_qsub}
\alias{submit_cmd_torque}
\alias{submit_job_sh}
\alias{submit_cmd_sh}
\alias{submit_cmd_local}
\title{Submit a script or command to a high-performance cluster using a scheduler}
\usage{
submit_job(
  scheduler = "slurm",
  input,
  is_script = TRUE,
  echo = TRUE,
  fail_on_error = FALSE,
  wait_jobs = NULL,
  env_variables = NULL,
  control = list()
)

submit_cmd(
  scheduler = "slurm",
  input,
  is_script = FALSE,
  echo = TRUE,
  fail_on_error = FALSE,
  wait_jobs = NULL,
  env_variables = NULL,
  control = list()
)

submit_job_sbatch(
  scheduler = "slurm",
  input,
  is_script = TRUE,
  echo = TRUE,
  fail_on_error = FALSE,
  wait_jobs = NULL,
  env_variables = NULL,
  control = list()
)

submit_job_slurm(
  scheduler = "slurm",
  input,
  is_script = TRUE,
  echo = TRUE,
  fail_on_error = FALSE,
  wait_jobs = NULL,
  env_variables = NULL,
  control = list()
)

submit_cmd_sbatch(
  scheduler = "slurm",
  input,
  is_script = FALSE,
  echo = TRUE,
  fail_on_error = FALSE,
  wait_jobs = NULL,
  env_variables = NULL,
  control = list()
)

submit_cmd_slurm(
  scheduler = "slurm",
  input,
  is_script = FALSE,
  echo = TRUE,
  fail_on_error = FALSE,
  wait_jobs = NULL,
  env_variables = NULL,
  control = list()
)

submit_job_qsub(
  scheduler = "torque",
  input,
  is_script = TRUE,
  echo = TRUE,
  fail_on_error = FALSE,
  wait_jobs = NULL,
  env_variables = NULL,
  control = list()
)

submit_job_torque(
  scheduler = "torque",
  input,
  is_script = TRUE,
  echo = TRUE,
  fail_on_error = FALSE,
  wait_jobs = NULL,
  env_variables = NULL,
  control = list()
)

submit_cmd_qsub(
  scheduler = "torque",
  input,
  is_script = FALSE,
  echo = TRUE,
  fail_on_error = FALSE,
  wait_jobs = NULL,
  env_variables = NULL,
  control = list()
)

submit_cmd_torque(
  scheduler = "torque",
  input,
  is_script = FALSE,
  echo = TRUE,
  fail_on_error = FALSE,
  wait_jobs = NULL,
  env_variables = NULL,
  control = list()
)

submit_job_sh(
  scheduler = "local",
  input,
  is_script = TRUE,
  echo = TRUE,
  fail_on_error = FALSE,
  wait_jobs = NULL,
  env_variables = NULL,
  control = list()
)

submit_cmd_sh(
  scheduler = "local",
  input,
  is_script = FALSE,
  echo = TRUE,
  fail_on_error = FALSE,
  wait_jobs = NULL,
  env_variables = NULL,
  control = list()
)

submit_cmd_local(
  scheduler = "local",
  input,
  is_script = FALSE,
  echo = TRUE,
  fail_on_error = FALSE,
  wait_jobs = NULL,
  env_variables = NULL,
  control = list()
)
}
\arguments{
\item{scheduler}{Which scheduler to use for job submission. Options are 'qsub',
'torque', 'sbatch', 'slurm', 'sh', or 'local'. The terms 'qsub' and 'torque'
are aliases (where 'torque' submits via the qsub command). Likewise for 'sbatch'
and 'slurm'. The scheduler 'sh' (alias, 'local') does not submit to any scheduler
at all, but instead executes the command immediately via sh. Default: 'slurm'}

\item{input}{A path to a script that should be executed by the scheduler or a
single command to be run. In the case of conflicts, the directives passed with
\code{sched_args} will take precedence. Required.}

\item{is_script}{Logical. If \code{TRUE}, assumes \code{input} is a script. If
false, assumes \code{input} is a single command. If NULL, the function will
internally determine what the input is. Default: \code{TRUE}}

\item{echo}{Whether to echo the job submission command to the terminal at the
time it is scheduled. Default: \code{TRUE}}

\item{fail_on_error}{Whether to stop execution of the script (\code{TRUE}), or issue
a warning (\code{FALSE}) if the job submission fails. Defaults: FALSE (i.e., issue a
warning).}

\item{wait_jobs}{An optional character string of jobs or process ids that should
complete before this job is executed}

\item{env_variables}{An optional named character vector containing environment variables
and their values to be passed to the \code{script} at execution time. This is
handled by the '-v' directive on TORQUE clusters and by '--export' on Slurm
clusters. The names of this vector are the environment variable names and the
values of the vector are the environment variable values to be passed in. If you
want to propagate the current value of an environment variable to the compute
node at runtime, use NA as the value of the element in \code{env_variables}.
See examples.}

\item{control}{A list of additional, scheduler-specific arguments. See Details.}
}
\value{
A character string containing the job ID of the scheduled job.
}
\description{
`submit_job` submits a single script or a single command to a high-performance
cluster (HPC) using a specified scheduler (TORQUE, Slurm) or locally. The function
`submit_cmd` is available to the user for submitting single commands
(`submit_job` assumes a script by default; see `is_script` argument). Also
available are scheduler-specific submit functions of the form `submit_job_<scheduler>`
or `submit_cmd_<scheduler>` where `<scheduler>` can be `slurm` or `sbatch` (aliases),
`torque` or `qsub` (aliases), or `local` or `sh` (aliases).
}
\details{
The \code{control} argument can include additional, scheduler-specific arguments
as a named list:
\describe{
   \item{\code{export_all}}{
   (For TORQUE or Slurm schedulers) Whether to export all environment variables
   to the compute node at runtime. Default: FALSE
   }
   \item{\code{sched_args}}{
   An optional character vector of arguments to be included in
   the scheduling command. On TORQUE, these will typically begin with '-l'
   such as '-l wall_time=10:00:00'.
   }
   \item{\code{wait_signal}}{
   (For TORQUE or Slurm schedulers) The signal that should
   indicate that parent jobs have finished. Default: 'afterok'
   }
   \item{\code{repolling_interval}}{
   (For local jobs) The number of seconds to wait
   before rechecking job status. Default: 60
   }
}
}
\examples{
\dontrun{
cmd <- 'squeue'
job_id <- submit_job(scheduler = "slurm", input = cmd, is_script = FALSE)}
}
\author{
Michael Hallquist, Zach Vig
}
