% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_job.R
\docType{methods}
\name{update_job}
\alias{update_job}
\title{Incrementally build a job using `+`}
\usage{
update_job(e1, e2, ...)
}
\arguments{
\item{e1}{A class_job object to be updated}

\item{e2}{A class_job_update object containing the updates}

\item{...}{Not currently used}

\item{warn_overwrite}{Whether to warn about overwritten properties}

\item{.call}{The calling environment for warning/error messages}

\item{use_default_settings}{Whether to use default validation settings}
}
\description{
Incrementally build a job using `+`
}
