% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_job.R
\docType{methods}
\name{update_job}
\alias{update_job}
\alias{+}
\title{Incrementally build a job using `+`}
\usage{
e1 + e2
}
\arguments{
\item{e1}{A class_job object to be updated}

\item{e2}{A class_job_update object containing the updates}
}
\description{
...
}
\details{
...
}
