% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wait_for_job.R
\name{wait_for_job}
\alias{wait_for_job}
\title{Pause execution of an R script while a scheduled job is not yet complete.}
\usage{
wait_for_job(
  job_ids,
  repolling_interval = 60L,
  max_wait = 60 * 60 * 24,
  scheduler = "local",
  quiet = TRUE,
  stop_on_timeout = TRUE
)
}
\arguments{
\item{job_ids}{One or more job ids of existing TORQUE or Slurm jobs, or process
ids of a local process for \code{scheduler = 'sh'}.}

\item{repolling_interval}{How often to recheck the job status, in seconds.
Default: 60}

\item{max_wait}{How long to wait on the job before giving up, in seconds.
Default: 24 hours (86,400 seconds)}

\item{scheduler}{What scheduler is used for job execution. Options are 'qsub'
and 'torque' (aliases), 'sbatch' and 'slurm' (aliases), and 'sh' or 'local'
(aliases). Default: 'local'}

\item{quiet}{Logical. If \code{TRUE}, \code{wait_for_job} will not print out
any status updates on jobs. If \code{FALSE}, the function prints out status
updates for each tracked job so that the user knows what's holding up progress.
Default: \code{TRUE}}

\item{stop_on_timeout}{Logical. If `TRUE`, the function throws an error if
the `max_wait` is exceeded. If `FALSE`, it returns `FALSE` instead of stopping.
Default: `TRUE`.}
}
\value{
Returns (invisibly) `TRUE` if all jobs completed successfully, `FALSE`
if any job failed or timeout occurred and `stop_on_timeout = FALSE`. Otherwise,
stops execution with an error if the timeout is exceeded.
}
\description{
`wait_for_job` gives you control over job dependencies within R when the formal
PBS depend approach is insufficient, especially in the case of a script that
spawns child jobs that need to be scheduled or complete before the parent script
should continue.
}
\details{
Note that for the \code{scheduler} argument, 'torque' and 'qsub' are aliases;
'slurm' and 'sbatch' are aliases, and 'sh' and 'local' are aliases.
}
\examples{
\dontrun{
wait_for_job(job_ids = "28465826", repolllig_interval = 60L,
             scheduler = "local", quiet = TRUE)}

}
\author{
Michael Hallquist, Zach Vig
}
